requirements.txt
pyyaml==5.1.2
prometheus_client==0.7.1
flask==1.1.1